Create assignment
Create a class called BankAccount that simulates a simple bank account. The class should have the following attributes and methods:

Attributes:
balance: a double representing the current balance in the account
accountNumber: an integer representing the account number
Methods:
BankAccount(double initialBalance, int accountNum): a constructor that initializes the balance and account number attributes

void deposit(double amount): a method that adds the specified amount to the account balance

void withdraw(double amount): a method that subtracts the specified amount from the account balance

double getBalance(): a method that returns the current account balance

int getAccountNumber(): a method that returns the account number Once you have defined the BankAccount class, create an instance of the class and perform the following operations on it:

Deposit $100 into the account

Withdraw $50 from the account

Print the current account balance and account number to the console

Note: You can assume that the user inputs the account number and initial balance when the BankAccount object is created.